few controller added

+++ server/controllers/Section.js
@@ -1,5 +1,7 @@
 const Section=require("../models/Section");
 const Course=require("../models/Course");
+const SubSection = require("../models/SubSection");
+
 
 //createSection
 exports.createSection=async(req,res)=>{
@@ -52,15 +54,25 @@ exports.createSection=async(req,res)=>{
 
 exports.updateSection=async(req,res)=>{
     try{
-        //data input
-        const{sectionName,sectionId}=req.body;
-        //update data
-        const section=await Section.findByIdAndUpdate(sectionId,{sectionName},{new:true})
-        //return res
-        return res.status(200).json({
-            success:true,
-            message:"section updated successfully!",
-            error:err.message
+        const { sectionName, sectionId, courseId } = req.body
+        const section = await Section.findByIdAndUpdate(
+          sectionId,
+          { sectionName },
+          { new: true }
+        )
+        const course = await Course.findById(courseId)
+          .populate({
+            path: "courseContent",
+            populate: {
+              path: "subSection",
+            },
+          })
+          .exec()
+        console.log(course)
+        res.status(200).json({
+          success: true,
+          message: section,
+          data: course,
         })
     }catch(err){
         return res.status(500).json({
@@ -73,24 +85,48 @@ exports.updateSection=async(req,res)=>{
 
 //deleteSection
 
-exports.deleteSection=async(req,res)=>{
-    try{
-        //get id--assuming that we are sending id in params
-        const {sectionId}=req.body
-        //delete it
-        await Section.findByIdAndDelete(sectionId);
-        //TODO: do we need to delete the entry from course schema???
-        //return res
-        return res.status(200).json({
-            success:true,
-            messagge:"Section Deleted Successfully!"
+exports.deleteSection = async (req, res) => {
+    try {
+      const { sectionId, courseId } = req.body
+      await Course.findByIdAndUpdate(courseId, {
+        $pull: {
+          courseContent: sectionId,
+        },
+      })
+      const section = await Section.findById(sectionId)
+      console.log(sectionId, courseId)
+      if (!section) {
+        return res.status(404).json({
+          success: false,
+          message: "Section not found",
         })
-    }catch(err){
-        return res.status(500).json({
-            success:false,
-            message:"Something went wrong in section updation",
-            error:err.message
+      }
+      // Delete the associated subsections
+      await SubSection.deleteMany({ _id: { $in: section.subSection } })
+  
+      await Section.findByIdAndDelete(sectionId)
+  
+      // find the updated course and return it
+      const course = await Course.findById(courseId)
+        .populate({
+          path: "courseContent",
+          populate: {
+            path: "subSection",
+          },
         })
+        .exec()
+  
+      res.status(200).json({
+        success: true,
+        message: "Section deleted",
+        data: course,
+      })
+    } catch (error) {
+      console.error("Error deleting section:", error)
+      res.status(500).json({
+        success: false,
+        message: "Internal server error",
+        error: error.message,
+      })
     }
-}
-
+  }
